Fix pulling people through space breaking the chain (#51997)
Includes a unit test because TDD is good.
Why It's Good For The Game

It makes transport through space with friends much more annoying than it should be.
Changelog

cl
fix: Fixed a bug where when teleporting naturally (such as going through a z-level in space), if you were pulling someone that was then pulling another person, only you and the person you were pulling would make it through.
/cl