HTML formatting on communication consoles, HTML sanitization (#54765)
Fixes #54713

Lets the comms console display HTML by telling react to set the HTML directly.

Also lets admins send HTML formatted command updates to the console. However, if the message is also sent to the crew, the raw HTML string will be sent in the chat box!

Shoutout to Timberpoes for helping me out with this!

Now with sanitize spun off into its own thing for anyone to reuse!

Accepts a string with HTML tags in it as input, and optionally allowed tags/forbidden attributes to override the defaults, then spits out a sanitised version of the same string!